refactor: restructure statistics components
