+++ app/(app)/statistics/_components/board/completed-questions.tsx
