+++ app/(app)/statistics/_components/board/index.tsx
