+++ app/(app)/statistics/_components/statistics/points.tsx
