+++ app/(app)/statistics/_components/statistics/resolved-questions.tsx
